Update XCast.cpp
type cast was needed to resolve the unit test failure

+++ XCast/XCast.cpp
@@ -963,11 +963,11 @@ void XCast::onXcastApplicationLaunchRequestWithLaunchParam (string appName,
     //TODO
     LOGINFO ("XcastService::onXcastApplicationLaunchRequestWithLaunchParam ");
     if(strAddDataUrl.size() > DIAL_MAX_ADDITIONALURL){
-        LOGWARN ("%s - current additional data size (%d) exceeds maximum allowed size (%d) ", __PRETTY_FUNCTION__, strAddDataUrl.size(), DIAL_MAX_ADDITIONALURL);
+        LOGWARN ("%s - current additional data size (%d) exceeds maximum allowed size (%d) ", __PRETTY_FUNCTION__,(int)strAddDataUrl.size(), DIAL_MAX_ADDITIONALURL);
         return;
     }
     if(strPayLoad.size() > DIAL_MAX_PAYLOAD) {
-        LOGWARN ("%s - current payload size (%d) exceeds maximum allowed size (%d) ", __PRETTY_FUNCTION__, strPayLoad.size(), DIAL_MAX_PAYLOAD);
+        LOGWARN ("%s - current payload size (%d) exceeds maximum allowed size (%d) ", __PRETTY_FUNCTION__,(int)strPayLoad.size(), DIAL_MAX_PAYLOAD);
         return;
     }
     JsonObject params;
